requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
asyncpg==0.30.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
redis==5.0.1
boto3==1.34.34
stripe==7.11.0
python-dotenv==1.0.0
jinja2==3.1.3
aiosmtplib==3.0.1
httpx==0.26.0
